+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
   * [tbc] LFG skin was broken
   * [classic] An extra texture was visible when at max level and not tracking any reputations
+  * The "At War" check box in the reputations frame was out of place
+  * The reputation detail frame had an out of place background
 
 
 ## [9.1.5.0] ##